+++ zmscitizenapi/src/Zmscitizenapi/AvailableAppointmentsList.php
@@ -5,22 +5,15 @@ namespace BO\Zmscitizenapi;
 use BO\Zmscitizenapi\BaseController;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
-use BO\Zmscitizenapi\Services\AvailableAppointmentsService;
+use BO\Zmscitizenapi\Services\ZmsApiFacadeService;
 
 class AvailableAppointmentsList extends BaseController
 {
-    protected $availableAppointmentsService;
-
-    public function __construct()
-    {
-        $this->availableAppointmentsService = new AvailableAppointmentsService();
-    }
-
     public function readResponse(RequestInterface $request, ResponseInterface $response, array $args)
     {
         $queryParams = $request->getQueryParams();
 
-        $result = $this->availableAppointmentsService->getAvailableAppointments($queryParams);
+        $result = ZmsApiFacadeService::getAvailableAppointments($queryParams);
 
         return $this->createJsonResponse($response, $result, $result['status']);
     }
