+++ zmscitizenapi/src/Zmscitizenapi/Helper/UtilityHelper.php
@@ -4,7 +4,7 @@ namespace BO\Zmscitizenapi\Helper;
 
 class UtilityHelper
 {
-    public function getInternalDateFromISO($dateString)
+    public static function getInternalDateFromISO($dateString)
     {
         $date = new \DateTime($dateString);
         return [
@@ -14,7 +14,7 @@ class UtilityHelper
         ];
     }
 
-    public function getInternalDateFromTimestamp(int $timestamp)
+    public static function getInternalDateFromTimestamp(int $timestamp)
     {
         $date = (new \DateTime())->setTimestamp($timestamp);
         return [
@@ -24,8 +24,56 @@ class UtilityHelper
         ];
     }
 
-    public function uniqueElementsFilter($value, $index, $self)
+    public static function uniqueElementsFilter($value, $index, $self)
     {
         return array_search($value, $self) === $index;
     }
+
+    public static function getThinnedProcessData($myProcess)
+    {
+        if (!$myProcess || !isset($myProcess->id)) {
+            return [];
+        }
+
+        $subRequestCounts = [];
+        $mainServiceId = null;
+        $mainServiceCount = 0;
+
+        if (isset($myProcess->requests)) {
+            $requests = is_array($myProcess->requests) ? $myProcess->requests : iterator_to_array($myProcess->requests);
+            if (count($requests) > 0) {
+                $mainServiceId = $requests[0]->id;
+                foreach ($requests as $request) {
+                    if ($request->id === $mainServiceId) {
+                        $mainServiceCount++;
+                    } else {
+                        if (!isset($subRequestCounts[$request->id])) {
+                            $subRequestCounts[$request->id] = [
+                                'id' => $request->id,
+                                'count' => 0,
+                            ];
+                        }
+                        $subRequestCounts[$request->id]['count']++;
+                    }
+                }
+            }
+        }
+
+        return [
+            'processId' => $myProcess->id,
+            'timestamp' => isset($myProcess->appointments[0]) ? $myProcess->appointments[0]->date : null,
+            'authKey' => $myProcess->authKey ?? null,
+            'familyName' => isset($myProcess->clients[0]) ? $myProcess->clients[0]->familyName : null,
+            'customTextfield' => $myProcess->customTextfield ?? null,
+            'email' => isset($myProcess->clients[0]) ? $myProcess->clients[0]->email : null,
+            'telephone' => isset($myProcess->clients[0]) ? $myProcess->clients[0]->telephone : null,
+            'officeName' => $myProcess->scope->contact->name ?? null,
+            'officeId' => $myProcess->scope->provider->id ?? null,
+            'scope' => $myProcess->scope ?? null,
+            'subRequestCounts' => array_values($subRequestCounts),
+            'serviceId' => $mainServiceId,
+            'serviceCount' => $mainServiceCount,
+        ];
+    }
+    
 }
