+++ zmscitizenapi/src/Zmscitizenapi/OfficesByServiceList.php
@@ -6,22 +6,16 @@ use BO\Zmscitizenapi\BaseController;
 use BO\Slim\Render;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
-use BO\Zmscitizenapi\Services\OfficesService;
+use BO\Zmscitizenapi\Services\ZmsApiFacadeService;
 
 class OfficesByServiceList extends BaseController
 {
-    protected $officesService;
-
-    public function __construct()
-    {
-        $this->officesService = new OfficesService();
-    }
 
     public function readResponse(RequestInterface $request, ResponseInterface $response, array $args)
     {
         $serviceIds = explode(',', $request->getQueryParams()['serviceId'] ?? '');
 
-        $result = $this->officesService->getOfficesByServiceIds($serviceIds);
+        $result = ZmsApiFacadeService::getOfficesByServiceIds($serviceIds);
 
         if (isset($result['error'])) {
             return $this->createJsonResponse($response, $result, $result['status']);
