+++ zmscitizenapi/src/Zmscitizenapi/OfficesServicesRelations.php
@@ -6,24 +6,14 @@ use BO\Zmscitizenapi\BaseController;
 use BO\Slim\Render;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
-use BO\Zmscitizenapi\Services\OfficesServicesRelationsService;
+use BO\Zmscitizenapi\Services\ZmsApiFacadeService;
 
 class OfficesServicesRelations extends BaseController
 {
-    protected $officesServicesRelationsService;
-
-    public function __construct()
-    {
-        $this->officesServicesRelationsService = new OfficesServicesRelationsService(); // No container
-    }
 
     public function readResponse(RequestInterface $request, ResponseInterface $response, array $args)
     {
-        $sources = \App::$http->readGetResult('/source/' . \App::$source_name . '/', [
-            'resolveReferences' => 2,
-        ])->getEntity();
-
-        $data = $this->officesServicesRelationsService->getOfficesServicesRelations($sources);
+        $data = ZmsApiFacadeService::getServicesAndOffices();
 
         return Render::withJson($response, $data);
     }
